requirements.txt
Django==5.1.7
djangorestframework==3.15.2
asgiref==3.8.1
sqlparse==0.5.3
langchain==0.3.20
langchain-deepseek==0.1.2
openai==1.65.5
python-dotenv==1.0.1 